requirements.txt
certifi==2018.11.29
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pycrypto==2.6.1
six==1.12.0
Werkzeug==0.14.1
wincertstore==0.2
